add GET / for endpoint

+++ src/restful/method/index.js
@@ -2,7 +2,7 @@ const debug = require('debug')('restful');
 
 const error = (e) => {
 	debug(e);
-	res.locals.error = 'An unexpected error occurred, it may cause by mongoose or the mongo server';
+	res.locals.error = e.message;
 	res.status(400);
 }
 
@@ -11,10 +11,12 @@ module.exports = (Model, name) => {
 
 	function get(req, res, next){
 		let id = res.locals.id;
+		let query = id ? { id } : {};
+		console.log({id, query});
 
-		Model.findOne({ id })
+		Model.find(query)
 			.then(data => {
-				res.locals.data = data;
+				res.locals.data = id ? data[0] : data;
 				res.status(200);
 			})
 			.catch(error)
@@ -24,6 +26,11 @@ module.exports = (Model, name) => {
 	function post(req, res, next){
 		let id = res.locals.id;
 		let data = req.data;
+
+		if(!id) {
+			error(new Error('id is undefined'));
+			return next();
+		}
 
 		Model.create({ id }, data)
 			.then(data => {
@@ -38,6 +45,11 @@ module.exports = (Model, name) => {
 		let id = res.locals.id;
 		let data = req.data;
 
+		if(!id) {
+			error(new Error('id is undefined'));
+			return next();
+		}
+		
 		Model.updateOne({ id }, data)
 			.then(data => {
 				res.locals.data = data;
@@ -51,6 +63,11 @@ module.exports = (Model, name) => {
 		let id = res.locals.id;
 		let data = req.data;
 
+		if(!id) {
+			error(new Error('id is undefined'));
+			return next();
+		}
+		
 		Model.deleteOne({ id }, data)
 			.then(data => {
 				res.locals.data = data;
